requirements.txt
notion_client==2.0.0
pandas==1.5.3
python-dotenv==1.0.0
ipython
jupyter